For 3D levels it's probably better if these dimension values can be floats

+++ Assets/Scripts/LevelEditor/LevelEditorData.cs
@@ -26,10 +26,10 @@ namespace R1Engine
         public static Unity_Level Level { get; set; }
         public static Unity_ObjectManager ObjManager => Level.ObjManager;
 
-        public static int MinX => 0;
-        public static int MinY => 0;
-        public static int MaxX => Level.MaxWidth;
-        public static int MaxY => Level.MaxHeight;
+        public static float MinX => 0;
+        public static float MinY => 0;
+        public static float MaxX => Level.MaxWidth;
+        public static float MaxY => Level.MaxHeight;
         public static float FramesPerSecond => Level?.FramesPerSecond ?? 60f;
 
         public static int CurrentLayer { get; set; }
